Optimise boot notification, set heartbeat to hourly
A charger can drop messages from the server until a valid boot notification response is received, so send acceptance before doing anything else. @lbbrhzn the heartbeat interval could be set in the config flow as the default for my charger is 12h or 43200 but others don't accept such a long interval.